refactor more of the summarize
