+++ waveanalysis/data_workflows/combined_workflow.py
@@ -12,7 +12,7 @@ from waveanalysis.image_props.image_bin_calc import create_multi_frame_bin_array
 from waveanalysis.image_props.image_to_np_arrays import tiff_to_np_array_multi_frame, tiff_to_np_array_single_frame
 from waveanalysis.image_props.image_properties import get_multi_frame_properties, get_single_frame_properties
 from waveanalysis.summarize_save.save_stats import save_parameter_means_to_csv, get_mean_CCF_values, get_indv_CCF_values, save_ccf_values_to_csv
-from waveanalysis.summarize_save.summarize_kymo_standard import summarize_image_standard_kymo, combine_stats_for_image_kymo_standard
+from waveanalysis.summarize_save.summarize_images import summarize_image, combine_stats_for_image_kymo_standard
 
 
 def combined_workflow(
@@ -326,7 +326,7 @@ def combined_workflow(
             ############################################
 
             # Summarize the data for current image as dataframe, and save as .csv
-            im_measurements_df, parameters_with_stats_dict = summarize_image_standard_kymo(
+            im_measurements_df, parameters_with_stats_dict = summarize_image(
                 img_parameters=img_parameters_dict,
                 img_props_dict=img_props_dict
             )
