+++ waveanalysis/data_workflows/rolling_workflow.py
@@ -13,7 +13,7 @@ import waveanalysis.housekeeping.housekeeping_functions as hf
 from waveanalysis.image_props.image_bin_calc import create_multi_frame_bin_array
 from waveanalysis.image_props.image_to_np_arrays import tiff_to_np_array_multi_frame
 from waveanalysis.image_props.image_properties import get_multi_frame_properties
-from waveanalysis.summarize_save.summarize_rolling import combine_stats_rolling, summarize_submovie_measurements
+from waveanalysis.summarize_save.summarize_images import summarize_image, combine_stats_rolling
 
 def rolling_workflow(
     folder_path: str,
@@ -223,9 +223,9 @@ def rolling_workflow(
             ############################################
 
             # summarize the data for each subframe as individual dataframes, and save as .csv
-            submovie_meas_list = summarize_submovie_measurements(
+            submovie_meas_list, _ = summarize_image(
                 img_props_dict=img_props_dict,
-                img_parameters_dict=img_parameters_dict
+                img_parameters=img_parameters_dict
             )
             csv_save_path = os.path.join(im_save_path, 'rolling_measurements')
             os.makedirs(csv_save_path, exist_ok=True)
