+++ waveanalysis/summarize_save/__init__.py
@@ -1,13 +1,11 @@
 from .save_stats import save_parameter_means_to_csv, get_mean_CCF_values, get_indv_CCF_values, save_ccf_values_to_csv
-from .summarize_kymo_standard import summarize_image_standard_kymo, combine_stats_for_image_kymo_standard
-from .summarize_rolling import combine_stats_rolling, summarize_submovie_measurements
+from .summarize_images import summarize_image, combine_stats_for_image_kymo_standard, combine_stats_rolling
 
 __all__ = [
     'save_parameter_means_to_csv',
-    'summarize_image_standard_kymo',
+    'summarize_image',
     'combine_stats_for_image_kymo_standard',
     'combine_stats_rolling',
-    'summarize_submovie_measurements',
     'get_mean_CCF_values',
     'get_indv_CCF_values',
     'save_ccf_values_to_csv'
